pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: BSD License",
]
name = "nomad-cau-plugin"
description = "Nomad plugin for CAU Kiel"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Stepanka Lankova", email = "stepanka.lankova@gmail.com" },
]
maintainers = [
    { name = "Stepanka Lankova", email = "stepanka.lankova@gmail.com" },
]
license = { file = "LICENSE" }
dependencies = [
    "nomad-lab>=1.3.0",
    "python-magic-bin; sys_platform == 'win32'",
    "pdfplumber>=0.10.0",
]

[project.urls]
Repository = "https://github.com/CAU-Kiel/nomad-cau-plugin"

[project.optional-dependencies]
dev = ["ruff", "pytest", "structlog"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4


[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # isort
    "I",
    # pylint
    "PL",
]

ignore = [
    "F403", # 'from module import *' used; unable to detect undefined names
]

fixable = ["ALL"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# this is entirely optional, you can remove this if you wish to
[tool.ruff.format]
# use single quotes for strings.
quote-style = "single"

# indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[project.entry-points.'nomad.plugin']
parser_entry_point = "nomad_cau_plugin.parsers:parser_entry_point"
schema_package_entry_point = "nomad_cau_plugin.schema_packages:schema_package_entry_point"
normalizer_entry_point = "nomad_cau_plugin.normalizers:normalizer_entry_point"
#measurement_schema = "nomad_cau_plugin.measurements:schema"
MRO005_schema = "nomad_cau_plugin.measurements:MRO005_schema"
MRO004_schema = "nomad_cau_plugin.measurements:MRO004_schema"

example_upload_entry_point = "nomad_cau_plugin.example_uploads:example_upload_entry_point"

[tool.setuptools_scm]
[tool.cruft]
# Avoid updating workflow files, this leads to permissions issues
skip = [".github/*"]

